requirements.txt
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==9.4.0